SpellCard activation logic
Still a work in progress. Within the SpellCard class: Added another constructor to take in TargetType as a parameter. Updated the Spells.csv to be compatible with it. Added an activate method and method for singletargets and for player targets. In the card class, I updated the useCard methods. Not sure if I did it correctly though.