GUI descriptors
Added some text regarding how to use the Spell Card when selected. Left some comments on the refresh() and refreshGUI() methods to indicate what needs to be reset when the player's turn ends such as the lblInstruction's text and the selectedTargets ArrayList. Didn't implement it into the playCard()'s conditionals in the case that the player wants to play multiple cards in one turn with the same selection.